refactor(settings/speech): Fixing wrong filename and cleaning mapDispatchToProps
