+++ src/components/Settings/Speech/index.js
@@ -1,1 +1,1 @@
-export { default } from './Speech';
+export { default } from './Speech.container';
